+++ src/main/java/model/datahandling/DataHandler.java
@@ -87,7 +87,7 @@ public class DataHandler {
     }
 
 
-    public static DateHashMap<Date, Float> getCurrencyData(String path) {
+    public static DateHashMap<Date, Double> getCurrencyData(String path) {
         IOException exception;
         try {
             return CurrencyExchangeReader.convertCSVFileToHandledData(path);
@@ -95,7 +95,7 @@ public class DataHandler {
             exception = e;
             System.out.println(e.getMessage());
         }
-        return new DateHashMap<>() {{ put(new Date(), 0f); }};
+        return new DateHashMap<>() {{ put(new Date(), 0d); }};
     }
 
     public static List<String> getMICs() {
